+++ src/features/editor/extensions.ts
@@ -20,7 +20,7 @@ import { Details } from "./extensions/details";
 import { DetailsContent } from "./extensions/details/content";
 import { DetailsSummary } from "./extensions/details/summary";
 import { Embed } from "./extensions/embed";
-import { EmbedPasteHandler } from "./extensions/embed/embedPasteHandler";
+import { EmbedPasteHandler } from "./extensions/embed/embed-paste-handler";
 import { SpeakerDeckEmbed } from "./extensions/embed/speacker-deck-embed";
 import Figure from "./extensions/figure";
 import { Caption } from "./extensions/figure/caption";
